// test/integration/test/ports/engines/queries/filters/one_relation/_fixture/common/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Blog: {
      readonly blogId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Comment: {
      readonly commentId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly likes: CodecTypes['pg/int4@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Post: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly popularity: CodecTypes['pg/int4@1']['output'];
      readonly blogId: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Blog: {
      readonly blogId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Comment: {
      readonly commentId: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly likes: CodecTypes['pg/int4@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Post: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly popularity: CodecTypes['pg/int4@1']['input'];
      readonly blogId: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly blog: {
      readonly blogId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly comment: {
      readonly commentId: CodecTypes['pg/text@1']['output'];
      readonly likes: CodecTypes['pg/int4@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'] | null;
      readonly text: CodecTypes['pg/text@1']['output'];
    };
    readonly post: {
      readonly blogId: CodecTypes['pg/text@1']['output'] | null;
      readonly popularity: CodecTypes['pg/int4@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly blog: {
      readonly blogId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly comment: {
      readonly commentId: CodecTypes['pg/text@1']['input'];
      readonly likes: CodecTypes['pg/int4@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'] | null;
      readonly text: CodecTypes['pg/text@1']['input'];
    };
    readonly post: {
      readonly blogId: CodecTypes['pg/text@1']['input'] | null;
      readonly popularity: CodecTypes['pg/int4@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;